# server/src/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: ReadySetHire API
  version: 1.0.0
  description: RESTful API for jobs, interviews, questions, applicants and answers.
servers:
  - url: /api
paths:
...
  /interviews/{interviewId}/applicants:
    get:
      summary: List applicants bound to interview
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: interviewId
          required: true
          schema: { type: integer }
      responses:
        '200': { description: OK }
    post:
      summary: Bind applicant to interview
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: interviewId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                applicant_id: { type: integer }
                status: { type: string }
              required: [applicant_id]
      responses:
        '200': { description: Bound }
  /interviews/{interviewId}/applicants/{applicantId}:
    patch:
      summary: Update applicant interview status
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: interviewId
          required: true
          schema: { type: integer }
        - in: path
          name: applicantId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status: { type: string }
      responses:
        '200': { description: Updated }
    delete:
      summary: Unbind applicant from interview
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: interviewId
          required: true
          schema: { type: integer }
        - in: path
          name: applicantId
          required: true
          schema: { type: integer }
      responses:
        '204': { description: Unbound }
  /applicants:
    get:
      summary: List applicants
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: OK }
    post:
      summary: Create applicant
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Applicant'
      responses:
        '201': { description: Created }
  /applicants/{id}:
    get:
      summary: Get applicant
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200': { description: OK }
    patch:
      summary: Update applicant
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Applicant'
      responses:
        '200': { description: Updated }
    delete:
      summary: Delete applicant
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '204': { description: Deleted }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Applicant:
      type: object
      properties:
        firstname: { type: string }
        surname: { type: string }
        emailAddress: { type: string }
        phoneNumber: { type: string }
      required: [firstname, surname, emailAddress]
...